requirements.txt
fastapi==0.111.0
boto3==1.34.87
google-cloud-storage==2.16.0
google-cloud-bigquery==3.19.0
uvicorn[standard]==0.29.0
python-json-logger==2.0.7
